Add header methods; Replace icap message separator with CRLF instead of LF
